// src/lib/generated/exhibition-media-manifest.ts
export const EXHIBITION_MEDIA_MANIFEST = {
  "Annihilation": {
    "sections": {
// ...
      "main": {
        "name": "main",
        "images": [
          {
            "name": "18.webp",
            "isWide": true
          }
        ]
      }
    }
  },
// ...
} as const;